Refactoring (at least 100 lines less in main Cartpole Class) and some small bugs regarding csv saving crushing for some settings.
