+++ CartPole/__init__.py
@@ -9,7 +9,6 @@ and many more. To run it needs some "environment": we provide you with GUI and d
 # Import module to save history of the simulation as csv file
 
 # Import module to interact with OS
-import os
 import traceback
 # Import module to get a current time and date used to name the files containing the history of simulations
 import timeit
@@ -17,6 +16,7 @@ import timeit
 
 import numpy as np
 import pandas as pd
+
 from Control_Toolkit.Controllers import template_controller
 from Control_Toolkit.others.environment import EnvironmentBatched
 from Control_Toolkit.others.globals_and_utils import (
@@ -26,8 +26,6 @@ from CartPole.cartpole_parameters import (J_fric, L, m_cart, M_fric, TrackHalfLe
                                           controlBias, controlDisturbance, CP_PARAMETERS_DEFAULT,
                                           g, k, m_pole, u_max, v_max, controlDisturbance_mode)
 # Interpolate function to create smooth random track
-from scipy.interpolate import BPoly, interp1d
-from scipy.stats import truncnorm
 # Run range() automatically adding progress bar in terminal
 from tqdm import trange
 
@@ -36,6 +34,8 @@ from CartPole._CartPole_mathematical_helpers import wrap_angle_rad
 from CartPole.latency_adder import LatencyAdder
 from CartPole.load import get_full_paths_to_csvs, load_csv_recording
 from CartPole.noise_adder import NoiseAdder
+from CartPole.noise_control_signal import add_control_noise
+from CartPole.random_target_generator import Generate_Random_Trace_Function
 from CartPole.state_utilities import (ANGLE_COS_IDX, ANGLE_IDX, ANGLE_SIN_IDX,
                                       ANGLED_IDX, POSITION_IDX, POSITIOND_IDX)
 from CartPole.state_utilities import create_cartpole_state
@@ -524,7 +524,7 @@ class CartPole(EnvironmentBatched):
                     {"target_position": self.target_position, "target_equilibrium": self.target_equilibrium, 'L': float(self.L_for_controller)}
                 ))
                 self.Q_update_time = timeit.default_timer()-update_start
-                self.Q_applied = add_control_noise(self.Q_calculated)
+                self.Q_applied = add_control_noise(self.Q_calculated, rng)
 
             self.Q = self.Q_applied
             self.dt_controller_steps_counter = 0
@@ -946,7 +946,7 @@ class CartPole(EnvironmentBatched):
                     self.time,
                     {"target_position": self.target_position, "target_equilibrium": self.target_equilibrium, "L": float(self.L_for_controller)}
                 ))
-                self.Q_applied = add_control_noise(self.Q_calculated)
+                self.Q_applied = add_control_noise(self.Q_calculated, rng)
 
 
             self.Q = self.Q_applied
@@ -1282,104 +1282,5 @@ class CartPole(EnvironmentBatched):
     # endregion
 
 
-# Generates a random target position
-# in a form of a function interpolating between turning points
-def Generate_Random_Trace_Function(
 
-        length_of_experiment,
-        rtf_rng,
 
-        track_relative_complexity,
-        interpolation_type,
-        turning_points,
-        turning_points_period,
-
-        start_random_target_position_at,
-        end_random_target_position_at,
-
-        used_track_fraction,
-):
-    if (turning_points is None) or (turning_points == []):
-
-        number_of_turning_points = int(np.floor(length_of_experiment * track_relative_complexity))
-
-        y = rtf_rng.uniform(-1.0, 1.0, number_of_turning_points)
-        y = y * used_track_fraction * TrackHalfLength
-
-        if number_of_turning_points == 0:
-            y = np.append(y, 0.0)
-            y = np.append(y, 0.0)
-        elif number_of_turning_points == 1:
-            if start_random_target_position_at is not None:
-                y[0] = start_random_target_position_at
-            elif end_random_target_position_at is not None:
-                y[0] = end_random_target_position_at
-            else:
-                pass
-            y = np.append(y, y[0])
-        else:
-            if start_random_target_position_at is not None:
-                y[0] = start_random_target_position_at
-            if end_random_target_position_at is not None:
-                y[-1] = end_random_target_position_at
-
-    else:
-        number_of_turning_points = len(turning_points)
-        if number_of_turning_points == 0:
-            raise ValueError('You should not be here!')
-        elif number_of_turning_points == 1:
-            y = np.array([turning_points[0], turning_points[0]])
-        else:
-            y = np.array(turning_points)
-
-    random_samples = number_of_turning_points - 2 if number_of_turning_points - 2 >= 0 else 0
-
-    if turning_points_period == 'random':
-        t_init = np.sort(rtf_rng.uniform(0.0, 1.0, random_samples))
-        t_init = np.insert(t_init, 0, 0.0)
-        t_init = np.append(t_init, 1.0)
-    elif turning_points_period == 'regular':
-        t_init = np.linspace(0, 1.0, num=random_samples + 2, endpoint=True)
-    else:
-        raise NotImplementedError('There is no mode corresponding to this value of turning_points_period variable')
-
-    t_init = t_init * length_of_experiment
-
-    # Try algorithm setting derivative to 0 a each point
-    if interpolation_type == '0-derivative-smooth':
-        yder = [[y[i], 0] for i in range(len(y))]
-        random_track_f = BPoly.from_derivatives(t_init, yder, extrapolate='periodic')
-    elif interpolation_type == 'linear':
-        random_track_f = interp1d(t_init, y, kind='linear', fill_value='extrapolate')
-    elif interpolation_type == 'previous':
-        random_track_f = interp1d(t_init, y, kind='previous', fill_value='extrapolate')
-    else:
-        raise ValueError('Unknown interpolation type.')
-
-    # Truncate the target position to be not grater than 80% of track length
-    def random_track_f_truncated(time):
-
-        target_position = random_track_f(time)
-        target_position = np.clip(target_position, -used_track_fraction * TrackHalfLength, used_track_fraction * TrackHalfLength)
-
-        return target_position
-
-    return random_track_f_truncated
-
-def add_control_noise(Q_calculated):
-
-    if controlDisturbance_mode == 'OFF':
-        Q_applied = Q_calculated
-    elif controlDisturbance_mode == 'additive':
-        Q_applied = Q_calculated + controlDisturbance * rng.standard_normal(
-            size=np.shape(Q_calculated), dtype=np.float32) + controlBias
-    elif controlDisturbance_mode == 'truncnorm':
-        scale = controlDisturbance
-        loc = Q_calculated + controlBias
-        Q_applied = truncnorm.rvs((-1.0 - loc) / scale, (1.0 - loc) / scale, loc=loc,
-                                  scale=scale, random_state=rng)
-        Q_applied = np.cast[np.float32](Q_applied)
-    else:
-        raise ValueError('controlDisturbance_mode with value {} not valid'.format(controlDisturbance_mode))
-
-    return Q_applied
